+++ src/main/java/com/github/lombrozo/testnames/rules/RuleNotContainsTestWord.java
@@ -21,75 +21,54 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.github.lombrozo.testnames.rules.ml;
+
+package com.github.lombrozo.testnames.rules;
 
 import com.github.lombrozo.testnames.Complaint;
 import com.github.lombrozo.testnames.Rule;
 import com.github.lombrozo.testnames.TestCase;
 import com.github.lombrozo.testnames.complaints.WrongTestName;
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.Locale;
-import java.util.regex.Pattern;
 import java.util.stream.Stream;
-import opennlp.tools.postag.POSTaggerME;
 
 /**
- * Present simple tense rule with using of ML.
+ * Test case without a 'test' word in name.
  *
- * @since 0.10
+ * @since 0.1.0
  */
-public final class PresentSimpleMlRule implements Rule {
+public final class RuleNotContainsTestWord implements Rule {
 
     /**
-     * The pattern to split camel case.
-     */
-    private static final Pattern CAMEL = Pattern.compile("(?=\\p{Lu})");
-
-    /**
-     * The Open NLP tagger.
-     */
-    private final POSTaggerME model;
-
-    /**
-     * The test to check.
+     * The test case.
      */
     private final TestCase test;
 
     /**
-     * The main constructor.
-     * @param tagger The Open NLP tagger
-     * @param tst The test to check
+     * Ctor.
+     *
+     * @param test The test case to check
      */
-    PresentSimpleMlRule(final POSTaggerME tagger, final TestCase tst) {
-        this.model = tagger;
-        this.test = tst;
+    RuleNotContainsTestWord(final TestCase test) {
+        this.test = test;
     }
 
     @Override
     public Collection<Complaint> complaints() {
-        final Tag[] parse = Tag.parse(
-            this.model.tag(
-                Stream
-                    .concat(
-                        Stream.of("It"),
-                        Arrays.stream(PresentSimpleMlRule.CAMEL.split(this.test.name()))
-                    ).map(s -> s.toLowerCase(Locale.ROOT))
-                    .toArray(String[]::new)
+        return new RuleConditional(
+            this::containsTest,
+            new WrongTestName(
+                this.test,
+                "test name doesn't have to contain the word 'test'"
             )
-        );
-        final Collection<Complaint> res;
-        if (parse[1].isVerb()) {
-            res = Collections.emptyList();
-        } else {
-            res = Collections.singleton(
-                new WrongTestName(
-                    this.test,
-                    "Test name should start with a verb"
-                )
-            );
-        }
-        return res;
+        ).complaints();
+    }
+
+    /**
+     * Is contains the 'test' word.
+     *
+     * @return The result
+     */
+    private boolean containsTest() {
+        return Stream.of("test", "TEST", "Test").anyMatch(this.test.name()::contains);
     }
 }
